+++ .gitignore
@@ -8,12 +8,13 @@ log.html
 /test.mli
 target
 Cargo.lock
+.build
 
 # These files would be generated by 'tree-sitter generate' with the default
 # settings. We don't want them because there's already a copy at the root.
-/interface/Cargo.toml
-/interface/binding.gyp
 /interface/bindings
-/ocaml/Cargo.toml
-/ocaml/binding.gyp
+/interface/binding.gyp
+/interface/Cargo.toml
 /ocaml/bindings
+/ocaml/binding.gyp
+/ocaml/Cargo.toml
