+++ Cargo.toml
@@ -12,13 +12,15 @@ edition = "2018"
 
 build = "bindings/rust/build.rs"
 include = [
-  "common/*",
-  "bindings/rust/*",
-  "ocaml/grammar.js",
-  "ocaml/src/*",
-  "interface/grammar.js",
-  "interface/src/*",
-  "queries/*"
+  "/bindings/rust/",
+  "/include/",
+  "/interface/src/node-types.json",
+  "/interface/src/parser.c",
+  "/interface/src/scanner.cc",
+  "/ocaml/src/node-types.json",
+  "/ocaml/src/parser.c",
+  "/ocaml/src/scanner.cc",
+  "/queries/"
 ]
 
 [lib]
