+++ bindings/rust/build.rs
@@ -1,10 +1,11 @@
 fn main() {
     let root_dir = std::path::Path::new(".");
+    let include_dir = root_dir.join("include");
     let ocaml_dir = root_dir.join("ocaml").join("src");
     let interface_dir = root_dir.join("interface").join("src");
 
     let mut c_config = cc::Build::new();
-    c_config.include(&ocaml_dir);
+    c_config.include(&include_dir);
     c_config
         .flag_if_supported("-Wno-unused-parameter")
         .flag_if_supported("-Wno-unused-but-set-variable")
@@ -12,7 +13,7 @@ fn main() {
 
     let mut cpp_config = cc::Build::new();
     cpp_config.cpp(true);
-    cpp_config.include(&ocaml_dir);
+    cpp_config.include(&include_dir);
     cpp_config
         .flag_if_supported("-Wno-unused-parameter")
         .flag_if_supported("-Wno-unused-but-set-variable");
